prepare mf-angular for publishing
