+++ libs/mf-angular/README.md
@@ -1,0 +1,6 @@
+# @mf/angular
+
+This library provides Angular Components for building Microfrontends
+
+
+
