+++ libs/mf-angular/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  displayName: 'ngx-mf-injector',
+  displayName: 'mf-angular',
   preset: '../../jest.preset.js',
   setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
   globals: {
@@ -14,7 +14,7 @@ module.exports = {
       },
     }
   },
-  coverageDirectory: '../../coverage/libs/ngx-mf-injector',
+  coverageDirectory: '../../coverage/libs/mf-angular',
   snapshotSerializers: [
     'jest-preset-angular/build/AngularNoNgAttributesSnapshotSerializer.js',
     'jest-preset-angular/build/AngularSnapshotSerializer.js',
