+++ libs/mf-angular/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/libs/ngx-mf-injector",
+  "dest": "../../dist/libs/mf-angular",
   "lib": {
     "entryFile": "src/index.ts"
   }
