+++ libs/mf-angular/package.json
@@ -1,14 +1,11 @@
 {
-  "name": "@valorkin/ngx-mf-injector",
-  "version": "0.0.1",
-  "description": "Module Federation Angular Container Injector for Microfrontend.",
-  "license": "Apache-2.0",
+  "name": "@mf/angular",
+  "version": "1.0.0",
+  "description": "Module Federation Angular container injector for Microfrontends",
+  "license": "MIT",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/valorkin/module-federation.git"
-  },
-  "publishConfig": {
-    "registry": "https://npm.pkg.github.com"
   },
   "engine": {
     "node": ">= 10"
@@ -16,7 +13,8 @@
   "peerDependencies": {
     "@angular/common": "ANGULAR_VER_PLACEHOLDER",
     "@angular/core": "ANGULAR_VER_PLACEHOLDER",
-    "@angular/elements": "ANGULAR_VER_PLACEHOLDER"
+    "@angular/elements": "ANGULAR_VER_PLACEHOLDER",
+    "@mf-core": "^1.0.0"
   },
   "dependencies": {
     "tslib": "^2.0.0"
