+++ libs/mf-angular/schematics/collection.json
