+++ libs/mf-angular/schematics/ng-add/schema.json
