+++ libs/mf-angular/schematics/ng-add/schema.ts
