+++ libs/mf-angular/schematics/ng-add/theme/sap_fiori_3_hcb.css
