+++ libs/mf-angular/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/ngx-container-injector.module'
+export * from './lib/components/public_api';
+export * from './lib/tokens';
