+++ libs/mf-angular/src/lib/components/container-injector/container-injector.component.scss
