+++ libs/mf-angular/src/lib/components/container-injector/public_api.ts
@@ -1,0 +1,3 @@
+export * from './container-injector.component';
+// export * from './plugin-page-launcher.component';
+export * from './container-injector.module';
