+++ libs/mf-angular/src/lib/components/public_api.ts
@@ -1,0 +1,3 @@
+export * from './container-injector/container-injector.module';
+export * from './element-container-injector/element-container-injector.module';
+export * from './container-injector/container-injector.component';
