+++ libs/mf-angular/tsconfig.lib.prod.json
