+++ libs/mf-angular/tsconfig.schematics.json
@@ -16,7 +16,7 @@
     "noUnusedParameters": true,
     "noUnusedLocals": true,
     "rootDir": "schematics",
-    "outDir": "../../dist/libs/ngx-mf-injector/schematics",
+    "outDir": "../../dist/libs/mf-angular/schematics",
     "skipDefaultLibCheck": true,
     "skipLibCheck": true,
     "sourceMap": true,
