--- libs/ngx-mf-injector/README.md
@@ -1,9 +1,0 @@
-# ngx-mf-injector
-
-This library was generated with [Nx](https://nx.dev).
-
-
-## Running unit tests
-
-Run `nx test ngx-mf-injector` to execute the unit tests.
-
