--- libs/ngx-mf-injector/src/index.ts
@@ -1,2 +1,0 @@
-
-            export * from './lib/sds';
