--- libs/ngx-mf-injector/src/lib/components/container-injector/plugin-launcher.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { PluginLauncherComponent } from './plugin-launcher.component';
-// import { PluginPageLauncherComponent } from './plugin-page-launcher.component';
-import { IframeErrorDirective } from '../../directives/iframe-error/iframe-error.directive';
-
-@NgModule({
-    imports: [
-        CommonModule
-    ],
-    declarations: [
-        PluginLauncherComponent,
-        // PluginPageLauncherComponent,
-        IframeErrorDirective
-    ],
-    exports: [
-        PluginLauncherComponent,
-        // PluginPageLauncherComponent
-    ]
-})
-export class PluginLauncherModule {
-}
