--- libs/ngx-mf-injector/src/lib/components/container-injector/public_api.ts
@@ -1,4 +1,0 @@
-export * from './plugin-launcher.component';
-// export * from './plugin-page-launcher.component';
-export * from './plugin-launcher.module';
-export * from './custom-elements-plugin-launcher.module';
