+++ package.json
@@ -13,7 +13,7 @@
     "ngcc": "ngcc",
     "build-mf-core": "nx run mf-core:build",
     "build-mf-rpc": "nx run mf-rpc:build:production",
-    "build-ngx-mf-injector": "nx run ngx-mf-injector:build:production",
+    "build-mf-angular": "nx run mf-angular:build:production",
     "//01": "commands to serve all the apps in parallel",
     "c-serve-01-shell": "npm start --prefix ./packages/one-bx-shell-app",
     "c-serve-02": "npm start --prefix ./packages/content-req-app",
@@ -48,6 +48,7 @@
     "@angular/router": "~11.2.9",
     "@fundamental-ngx/app-shell": "^0.2.38",
     "@fundamental-ngx/core": "0.28.1",
+    "@mf/core": "^1.0.0",
     "@pscoped/ngx-pub-sub": "^3.0.0",
     "ngx-build-plus": "^11.0.0",
     "rxjs": "~6.5.4",
@@ -67,6 +68,7 @@
     "@nrwl/eslint-plugin-nx": "12.3.4",
     "@nrwl/jest": "11.2.9",
     "@nrwl/linter": "12.3.4",
+    "@nrwl/node": "11.2.9",
     "@nrwl/tao": "latest",
     "@nrwl/workspace": "latest",
     "@types/jasmine": "~3.6.0",
